+++ tests/testthat/test-session.R
@@ -171,7 +171,6 @@ test_that("pkg_version_info() parses info", {
   ))
 
   expect_snapshot({
-    (expect_error(pkg_version_info("foo 1.0"), "valid"))
     (expect_error(pkg_version_info("foo (1.0)"), "parse"))
     (expect_error(pkg_version_info("foo (>= 1.0)", "1.0"), "both"))
     (expect_error(pkg_version_info(c("foo (!= 1.0)"))))
@@ -233,4 +232,10 @@ test_that("is_installed('base') works (#1434)", {
     expect_true(is_installed(pkg, version = r_ver))
     expect_false(is_installed(pkg, version = "999.9.9"))
   }
+})
+
+test_that("is_installed() allows irregular package names", {
+  # Consistently with `loadNamespace()`
+  expect_false(is_installed("foo 1.0"))
+  expect_false(is_installed("::", version = "1.0"))
 })